+++ openAI_RRM/channel_controller.py
@@ -9,6 +9,8 @@ from uniflex.core.timer import TimerEventSender
 from common import AveragedSpectrumScanSampleEvent
 from common import ChangeWindowSizeEvent
 
+from UniFlexGym.interfaces.uniflex_controller import UniFlexController
+
 __author__ = "Piotr Gawlowicz, Sascha Rösler"
 __copyright__ = "Copyright (c) 2016, Technische Universität Berlin"
 __version__ = "0.1.0"
@@ -20,7 +22,7 @@ class PeriodicEvaluationTimeEvent(events.TimeEvent):
         super().__init__()
 
 
-class UniflexChannelController(modules.ControlApplication):
+class UniflexChannelController(modules.ControlApplication, UniFlexController):
     def __init__(self,**kwargs):
         super(UniflexChannelController, self).__init__()
         self.log = logging.getLogger('ChannelController')
@@ -333,7 +335,104 @@ class UniflexChannelController(modules.ControlApplication):
                     self.channel += 1
                     if self.channel > 13:
                         self.channel = 1
+    
+    '''
+    OpenAI Gym Uniflex env API
+    '''
+    
+    def __init__(self, **kwargs):
+        super()
+        self.observationSpace = []
+        self.lastObservation = []
+        return
+    
+    def reset(self):
+        self.observationSpace = self._create_client_list()
+        self.actionSpace = self._create_interface_list()
         
+        interfaces = self.get_interfaces()
+        
+        # set a start channel for each interface:
+        channel = 1
+        for node in interfaces:
+            for device in node['devices']:
+                for iface in device['interfaces']:
+                    self.set_channel(
+                        node['uuid'], device['uuid'], iface, channel, None)
+                    channel += 5
+                    if channel > 12:
+                        channel = 1
+        # clear bandwidth counter
+        self.get_bandwidth()
+        return
+    
+    def execute_action(self, action):
+        for index, actionStep in action:
+            interface = self.actionSpace[index]
+            self.set_channel(interface['node'], interface['device'], interface['iface'], actionStep, None)
+        return
+    
+    def render():
+        return
+    
+    def get_observationSpace(self):
+        return
+    
+    def get_actionSpace(self):
+        return
+    
+    def get_observation(self):
+        observation  = []
+        bandwidth = self.get_bandwidth()
+        bandwidth = sorted(bandwidth, key=lambda k: k['mac'])
+        for client in self.observationSpace:
+            bandwidth = self. _get_bandwidth_by_client( bandwidth, client)
+            if bandwidth in None:
+                bandwidth = 0
+            observation.append(bandwidth)
+        
+        self.lastObservation = observation
+        return observation
+    
+    # game over if there is a new interface
+    def get_gameOver(self):
+        clients = self._create_client_list()
+        return len(set(clients).symmetric_difference(set(self.observationSpace))) == 0
+    
+    def get_reward(self):
+        if len(self.lastObservation) > 0:
+            return reduce(lambda x, y: x^2 + y, self.lastObservation)
+        return 0
+    
+    
+    
+    def _get_bandwidth_by_client(self, bandwidthList, clientData):
+        for client in bandwidthList:
+            if (client['mac'] is clientData['mac']) and (client['node'] is clientData['node']) and (client['device'] is clientData['device']) and (client['iface'] is clientData['iface']):
+                return client['bandwidth']
+        return None
+    
+    def _create_client_list(self):
+        clientList = []
+        clients = self.get_bandwidth()
+        for client in clients:
+            clientList.append({'mac': client['mac'], 'node': client['node']['uuid'],
+                'device': client['device']['uuid'], 'iface': client['interface']})
+        clients = sorted(clients, key=lambda k: k['mac'])
+        return clientList
+    
+    def _create_interface_list(self):
+        interfaceList = []
+        interfaces = self.get_interfaces()
+        for node in interfaces:
+            for device in node['devices']:
+                for iface in device['interfaces']:
+                    interfaceList.append({'node': node['uuid'], 'device': device['uuid'], 'iface': iface})
+        return interfaceList
+    
+    
+    
+    
         '''
         print(self.get_bandwidth())
